requirements.txt
azure-functions
httpx
psycopg2-binary
python-dotenv
